backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
httpx==0.28.1
pillow>=12.1.1
python-multipart>=0.0.18
slowapi==0.1.9
pydantic==2.9.0
stripe==11.3.0
openai>=1.0.0
python-dotenv==1.0.0
PyJWT>=2.10.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
